# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateAdminRequest:
    properties:
      department:
        example: Administration
        type: string
      email:
        example: admin2@example.com
        type: string
      firstname:
        example: Admin
        type: string
      lastname:
        example: User
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: admin2
        type: string
    required:
    - email
    - firstname
    - lastname
    - password
    - username
    type: object
  handlers.CreateEmployeeRequest:
    properties:
      department:
        example: HR
        type: string
      email:
        example: jane@example.com
        type: string
      firstname:
        example: Jane
        type: string
      lastname:
        example: Smith
        type: string
      nrc:
        example: 555666/77/8
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        example: manager
    required:
    - email
    - firstname
    - lastname
    - nrc
    - password
    - role
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid credentials
        type: string
    type: object
...
host: localhost:8070
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    REST API for Leave Management System - part of HRMS
    Allows employees to apply for leaves, track balances, and managers to approve/reject requests.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: HRMS Leave Management API
  version: "1.0"
paths:
  /api/admins:
    post:
      consumes:
      - application/json
      description: Create a new admin account with username (Admin only)
      parameters:
      - description: Admin data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Username or email already exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create admin
      tags:
      - Admin - Employees
  /api/audit-logs:
    get:
      description: Get audit logs with optional filtering by entity type, entity ID,
        or performed by
      parameters:
      - description: Entity type filter
        in: query
        name: entity_type
        type: string
      - description: Entity ID filter
        in: query
        name: entity_id
        type: integer
      - description: Performed by user ID filter
        in: query
        name: performed_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audit logs
      tags:
      - Core HR - Audit
  /api/compliance/requirements:
    get:
      description: Get list of all active compliance requirements
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ComplianceRequirement'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all compliance requirements
      tags:
      - Core HR - Compliance
    post:
      consumes:
      - application/json
      description: Create a new compliance requirement (Manager/Admin only)
      parameters:
      - description: Compliance requirement data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ComplianceRequirement'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ComplianceRequirement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create compliance requirement
      tags:
      - Core HR - Compliance
  /api/employees:
    get:
      description: Get list of all employees (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all employees
      tags:
      - Admin - Employees
    post:
      consumes:
      - application/json
      description: Create a new employee or manager account with NRC (Admin only).
        Use /api/admins for admin accounts.
      parameters:
      - description: Employee data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: NRC or email already exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create employee/manager
      tags:
      - Admin - Employees
  /api/employees/{id}:
    delete:
      description: Delete an employee (Admin only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete employee
      tags:
      - Admin - Employees
    get:
      description: Get a specific employee by ID (Admin only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee by ID
      tags:
      - Admin - Employees
    put:
      consumes:
      - application/json
      description: Update an employee's information (Admin only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update employee
      tags:
      - Admin - Employees
  /api/employees/{id}/audit-logs:
    get:
      description: Get audit logs related to a specific employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee audit logs
      tags:
      - Core HR - Audit
  /api/employees/{id}/compliance:
    get:
      description: Get all compliance records for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ComplianceRecord'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee compliance records
      tags:
      - Core HR - Compliance
    post:
      consumes:
      - application/json
      description: Create a new compliance record for an employee (Manager/Admin only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Compliance record data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ComplianceRecord'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ComplianceRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create compliance record
      tags:
      - Core HR - Compliance
  /api/employees/{id}/documents:
    get:
      description: Get all documents for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Document'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee documents
      tags:
      - Core HR - Documents
    post:
      consumes:
      - multipart/form-data
      description: Upload a file and create a new document record for an employee.
        Accepts multipart/form-data with file upload.
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Document file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Document type (id, contract, resume, certificate, license, performance,
          disciplinary, compliance, other)
        in: formData
        name: document_type
        required: true
        type: string
      - description: Document title
        in: formData
        name: title
        required: true
        type: string
      - description: Document description
        in: formData
        name: description
        type: string
      - description: Issue date (YYYY-MM-DD)
        in: formData
        name: issue_date
        type: string
      - description: Expiry date (YYYY-MM-DD)
        in: formData
        name: expiry_date
        type: string
      - description: Is document confidential
        in: formData
        name: is_confidential
        type: boolean
      - description: Document tags (comma-separated)
        in: formData
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "415":
          description: Unsupported file type
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload and create document
      tags:
      - Core HR - Documents
  /api/employees/{id}/documents/{doc_id}:
    delete:
      description: Delete a document record and its associated file
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - Core HR - Documents
  /api/employees/{id}/documents/{doc_id}/download:
    get:
      description: Download the actual file for a document
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download document file
      tags:
      - Core HR - Documents
  /api/employees/{id}/employment:
    get:
      description: Get employment details for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmploymentDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee employment details
      tags:
      - Core HR - Employment
    post:
      consumes:
      - application/json
      description: Create or update employment details for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EmploymentDetails'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmploymentDetails'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EmploymentDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create or update employment details
      tags:
      - Core HR - Employment
  /api/employees/{id}/employment/history:
    get:
      description: Get employment history for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EmploymentHistory'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee employment history
      tags:
      - Core HR - Employment
  /api/employees/{id}/identity:
    get:
      description: Get identity information for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdentityInformation'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee identity information
      tags:
      - Core HR - Identity
    post:
      consumes:
      - application/json
      description: Create or update identity information for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Identity information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.IdentityInformation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdentityInformation'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdentityInformation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create or update identity information
      tags:
      - Core HR - Identity
  /api/employees/{id}/lifecycle:
    get:
      description: Get all lifecycle events for an employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.WorkLifecycleEvent'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee lifecycle events
      tags:
      - Core HR - Lifecycle
    post:
      consumes:
      - application/json
      description: Create a new lifecycle event for an employee (Manager/Admin only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lifecycle event data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WorkLifecycleEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.WorkLifecycleEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create lifecycle event
      tags:
      - Core HR - Lifecycle
...
  /api/employees/bulk:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple employees from a CSV file (Admin only)
      parameters:
      - description: CSV file with employee data
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BulkUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk upload employees
      tags:
      - Admin - Employees
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (without quotes)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"